Add sed to fix broken emails in bzr
Some emails in a launchpad project I use (lp:graphite) have no
username. Without this fix, they are imported as "user@example.com <>"
which is not a valid format in git.

This fix copies the username out of the email and forms a new id like:
"user <user@example.com>".